tests/krate/publish: Extract `inheritance` module
